# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Keeppix API",
    "description": "API di Keeppix. Contratto congelato: solo aggiunte entro /api/v1.",
    "license": {
      "name": "AGPL-3.0-or-later",
      "identifier": "AGPL-3.0-or-later"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/assets/{id}": {
      "get": {
        "tags": [
          "timeline"
        ],
        "summary": "Get an asset",
        "operationId": "assets_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id asset",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vista pubblica dell'asset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetView"
                }
              }
            }
          },
          "400": {
            "description": "Id non valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Asset non visibile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Asset inesistente per un admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      },
      "delete": {
        "tags": [
          "trash"
        ],
        "summary": "Delete an asset",
        "operationId": "assets_delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id dell'asset",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Azione applicata, riga di audit registrata"
          },
          "400": {
            "description": "disk_action non riconosciuto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Asset non visibile, o purged senza i permessi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Errore del database o del filesystem",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
    "/api/v1/duplicates": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "List duplicate groups",
        "operationId": "duplicates_list",
        "responses": {
          "200": {
            "description": "Gruppi con lo stesso content_hash, esclusi i trashed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DuplicateGroupView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/duplicates/{content_hash}": {
      "get": {
        "tags": [
          "library"
        ],
        "summary": "List members of a duplicate group",
        "operationId": "duplicates_members",
        "parameters": [
          {
            "name": "content_hash",
            "in": "path",
            "description": "blake3 hex da 64 caratteri",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Membri non cestinati del gruppo",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetView"
                  }
                }
              }
            }
          },
          "400": {
            "description": "content_hash non è hex valido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
    "/api/v1/duplicates/{content_hash}/resolve": {
      "post": {
        "tags": [
          "library"
        ],
        "summary": "Resolve a duplicate group",
        "operationId": "duplicates_resolve",
        "parameters": [
          {
            "name": "content_hash",
            "in": "path",
            "description": "blake3 hex da 64 caratteri",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveDuplicateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Azione applicata agli altri membri",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveDuplicateResponse"
                }
              }
            }
          },
          "400": {
            "description": "content_hash o disk_action non validi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Non autenticato",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "keep non nel gruppo, o purged senza i permessi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetView": {
        "type": "object",
        "required": [
          "id",
          "folder_id",
          "filename",
          "size_bytes",
          "kind",
          "status"
        ],
        "properties": {
          "content_hash": {
            "type": [
              "string",
              "null"
            ]
          },
          "filename": {
            "type": "string"
          },
          "folder_id": {
            "type": "string"
          },
          "height": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "location": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/GeoPointView"
              }
            ]
          },
          "place_id": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "taken_at_utc": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "thumbhash": {
            "type": [
              "string",
              "null"
            ]
          },
          "width": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          }
        }
      },
...
      "DeleteAssetRequest": {
        "type": "object",
        "required": [
          "disk_action"
        ],
        "properties": {
          "disk_action": {
            "type": "string",
            "description": "`kept`, `moved_to_trash`, o `purged` — nessun default: il client deve\nsempre scegliere (spec §6).",
            "example": "moved_to_trash"
          }
        }
      },
...
      "DuplicateGroupView": {
        "type": "object",
        "required": [
          "content_hash",
          "count",
          "size_bytes",
          "reclaimable_bytes"
        ],
        "properties": {
          "content_hash": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "reclaimable_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "GeoPointView": {
        "type": "object",
        "required": [
          "lat",
          "lon"
        ],
        "properties": {
          "lat": {
            "type": "number",
            "format": "double"
          },
          "lon": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "Problem": {
        "type": "object",
        "description": "Errore in formato RFC 9457. Il campo `type` è un codice stabile su cui i\nclient possono ramificare; `title` è in inglese e serve al debug, non\nall'utente finale — la traduzione avviene nel frontend.\n\nÈ anche uno schema del documento `OpenAPI`: è la forma su cui il client\nmobile ramifica, quindi va descritta, non solo servita.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "example": "keeppix/unauthenticated"
          }
        }
      },
...
      "ResolveDuplicateRequest": {
        "type": "object",
        "required": [
          "keep",
          "disk_action"
        ],
        "properties": {
          "disk_action": {
            "type": "string",
            "description": "`kept`, `moved_to_trash`, o `purged` — stesso vocabolario di\n`DELETE /api/v1/assets/{id}` (spec §6).",
            "example": "moved_to_trash"
          },
          "keep": {
            "type": "string",
            "description": "Id dell'asset da tenere: tutti gli altri membri non cestinati del\ngruppo ricevono `disk_action`."
          }
        }
      },
      "ResolveDuplicateResponse": {
        "type": "object",
        "required": [
          "resolved"
        ],
        "properties": {
          "resolved": {
            "type": "integer",
            "description": "Quanti membri sono stati toccati (il gruppo meno `keep`).",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-kpx_session",
        "description": "Cookie di sessione emesso da POST /api/v1/setup e POST /api/v1/auth/login."
      }
    }
  },
...
}